Adding generate.pl, a perl script I've written to generate static blog pages. It is commented so I won't go into too much detail here, but basically it looks through the year directories and looks for folders that include an article HTML file with the same name as the folder, and a title.txt file. Folders that meet those requirements have their index.html file overwritten with a generated blog page using a template.
The idea here is to allow for this blog to remain static and homegrown, but
make it easier for me to adjust the HTML and CSS surrounding old blog posts.
Instead of hand-editing each blog post to use a new profile picture, for
example, I'll be able to edit the template file and then apply that change
to all posts with a single command.

I'm sure there will be kinks I'll have to work out here, and there are
certainly improvements I'm already thinking of making. It would be cool for
the script to automatically overwrite the /articles.html page with all the
posts by year, and it would also be cool if I could have it generate the
homepage to include whichever articles I want.